restructure folders to bulletproof react specs
