+++ src/components/Post/Collaborator.tsx
