+++ src/components/Post/ProposalTag.tsx
